+++ src/App.vue
@@ -64,54 +64,6 @@ export default {
       return this.Panelmovie.title.toLowerCase();
     }
   },
-  mounted() {
-    dragElement(document.getElementById("keyboard"));
-
-    function dragElement(elmnt) {
-      var pos1 = 0,
-        pos2 = 0,
-        pos3 = 0,
-        pos4 = 0;
-      if (document.getElementById(elmnt.id + "header")) {
-        /* if present, the header is where you move the DIV from:*/
-        document.getElementById(elmnt.id + "header").onmousedown =
-          dragMouseDown;
-      } else {
-        /* otherwise, move the DIV from anywhere inside the DIV:*/
-        elmnt.onmousedown = dragMouseDown;
-      }
-
-      function dragMouseDown(e) {
-        e = e || window.event;
-        e.preventDefault();
-        // get the mouse cursor position at startup:
-        pos3 = e.clientX;
-        pos4 = e.clientY;
-        document.onmouseup = closeDragElement;
-        // call a function whenever the cursor moves:
-        document.onmousemove = elementDrag;
-      }
-
-      function elementDrag(e) {
-        e = e || window.event;
-        e.preventDefault();
-        // calculate the new cursor position:
-        pos1 = pos3 - e.clientX;
-        pos2 = pos4 - e.clientY;
-        pos3 = e.clientX;
-        pos4 = e.clientY;
-        // set the element's new position:
-        elmnt.style.top = elmnt.offsetTop - pos2 + "px";
-        elmnt.style.left = elmnt.offsetLeft - pos1 + "px";
-      }
-
-      function closeDragElement() {
-        /* stop moving when mouse button is released:*/
-        document.onmouseup = null;
-        document.onmousemove = null;
-      }
-    }
-  },
   methods: {
     // Obtiene los datos de la aplicación
     async getData() {
@@ -275,7 +227,7 @@ export default {
       Volver a jugar
     </button>
     <panel-letters :text="movieTitle" :guessedLetters="guessedLetters" />
-    <keyboard id="keyboard" v-if="gameStatus == 0" :popcornNumber="tryNumber" :letters="letterArray"
+    <keyboard id="keyboard" v-if="gameStatus == 0 && actualMovie.length!=0" :popcornNumber="tryNumber" :letters="letterArray"
       @clickedLetter="(id) => letterClicked(id)" />
   </main>
 </template>
@@ -303,6 +255,7 @@ main #app {
 header {
   line-height: 1.5;
 }
+
 
 .logo {
   width: 80px;
@@ -367,9 +320,4 @@ a,
   }
 }
 
-#keyboard {
-  position: absolute;
-  left: 50%;
-  transform: translate(-50%);
-}
 </style>
