+++ src/components/Keyboard.vue
@@ -1,4 +1,5 @@
-
+<script setup>
+import { dragElement } from "@/utils/utils.js";</script>
 <script>
 export default {
   props: {
@@ -14,6 +15,9 @@ export default {
       this.$emit("clickedLetter", letter);
     },
   },
+  mounted() {
+    dragElement(document.getElementById("keyboardContainer"),document.getElementById("drag-icon"));    
+  }
 };
 </script>
 
@@ -27,23 +31,27 @@ export default {
         </div>
       </div>
     </div>
-    <p class="keyboardLines" v-for="(letterRow, index) in letters" :key="index">
-      <span
-        v-for="(letter, index2) in letterRow"
-        :key="index2"
-        @click="checkLetter(letter.letter)"
-        class="keyStyles"
-        :class="
-          letter.status == 'default'
-            ? 'keyIsDefault'
-            : letter.status == 'correct'
-            ? 'keyIsCorrect'
-            : 'keyIsWrong'
-        "
-      >
-        {{ letter.letter }}
-      </span>
-    </p>
+    
+    <div id="keyboardContainer">
+      <span id="drag-icon"><i class="fa-solid fa-arrows-up-down-left-right"></i></span>
+      <p class="keyboardLines" v-for="(letterRow, index) in letters" :key="index">
+        <span
+          v-for="(letter, index2) in letterRow"
+          :key="index2"
+          @click="checkLetter(letter.letter)"
+          class="keyStyles"
+          :class="
+            letter.status == 'default'
+              ? 'keyIsDefault'
+              : letter.status == 'correct'
+              ? 'keyIsCorrect'
+              : 'keyIsWrong'
+          "
+        >
+          {{ letter.letter }}
+        </span>
+      </p>
+    </div>
   </div>
 </template>
 
@@ -52,10 +60,11 @@ export default {
 div {
   padding: 10px;
 }
-.keyboardLines {
-  text-align: center;
-}
 
+#drag-icon{
+  font-size: 40px;
+  cursor: grab;
+}
 .keyStyles {
   background-color: #303030 ;
   text-align: center;
@@ -110,6 +119,22 @@ div {
     display: flex;
     align-items: center;
     justify-content: center;
+}
+
+#keyboardContainer {
+  position: absolute;
+  left: 50%;
+  transform: translate(-50%);
+}
+  .keyboardLines{
+    display:flex;
+    flex-direction: row;
+    justify-content: center;
+
+  }
+.keyboardLines, #keyboardContainer {
+  text-align: center;
+  flex-wrap: nowrap !important;
 }
 
 </style>
